// public-ui/src/publicGenerated/model/dbDomain.ts
/**
 * AllOfUs Public API
 * The API for the AllOfUs data browser and public storefront.
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface DbDomain {
    /**
     * domain id
     */
    domainId: string;

    /**
     * web display name
     */
    domainDisplay?: string;

    /**
     * full desc text
     */
    domainDesc?: string;

    /**
     * type of domain
     */
    dbType?: string;

    /**
     * route
     */
    domainRoute?: string;

    /**
     * concept id of domain
     */
    conceptId?: number;

    /**
     * count value
     */
    countValue?: number;

}