// public-ui/src/publicGenerated/model/questionConcept.ts
/**
 * AllOfUs Public API
 * The API for the AllOfUs data browser and public storefront.
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Analysis } from './analysis';


export interface QuestionConcept {
    /**
     * id of the concept
     */
    conceptId: number;

    /**
     * name of concept
     */
    conceptName: string;

    /**
     * domain of concept
     */
    domainId: string;

    /**
     * original vocab code of concept
     */
    conceptCode: string;

    /**
     * est count in the cdr
     */
    countValue?: number;

    /**
     * prevalence among participants percent count divided num participants
     */
    prevalence?: number;

    /**
     * count analysis
     */
    countAnalysis?: Analysis;

    /**
     * gender analysis
     */
    genderAnalysis?: Analysis;

    /**
     * age analysis
     */
    ageAnalysis?: Analysis;

}