// public-ui/src/publicGenerated/model/questionConceptListResponse.ts
/**
 * AllOfUs Public API
 * The API for the AllOfUs data browser and public storefront.
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DbDomain } from './dbDomain';
import { QuestionConcept } from './questionConcept';


export interface QuestionConceptListResponse {
    survey?: DbDomain;

    items: Array<QuestionConcept>;

}